+++ WHATSNEW.md
@@ -10,7 +10,8 @@ This document attempts to track **major** changes and additions in ENiGMA½. For
   * Setting the `data` member of an object will cause form submissions to use this value instead of the selected items index.
   * See the default `luciano_blocktronics` `matrix` menu for example usage.
 * You can now set the `sort` property on a menu to sort items. If `true` items are sorted by `text`. If the value is a string, it represents the key in menu objects to sort by.
-
+* Hot-reload of configuration files such as menu.hjson, config.hjson, your themes.hjson, etc.: When a file is saved, it will be hot-reloaded into the running system
+  * Note that any custom modules should make use of the new Config.get() method.
 
 ## 0.0.8-alpha
 * [Mystic BBS style](http://wiki.mysticbbs.com/doku.php?id=displaycodes) extended pipe color codes. These allow for example, to set "iCE" background colors.